EC2_dev/requirement.txt
selenium
aspose-words
Pillow
boto3
mysql-connector-python
pandas